+++ lib/base/index.js
@@ -1,22 +1,24 @@
 // @flow
 import {keybaseExec, rmdirRecursive, randomTempDir, keybaseServiceStartup} from '../utils'
 
+export type Initialized = false | 'paperkey' | 'currentService'
+
 export type InitOptions = {|
   verbose?: boolean,
+  homeDir?: string,
 |}
 
 export type DeviceUsernamePair = {|username: string, devicename: string|}
 
 class BaseBot {
-  initialized: boolean
+  initialized: Initialized
   username: void | string
   devicename: void | string
-  homeDir: string
+  homeDir: void | string
   verbose: boolean
 
   constructor() {
     this.initialized = false
-    this.homeDir = randomTempDir()
     this.verbose = false
   }
 
@@ -29,15 +31,18 @@ class BaseBot {
       throw new Error(`Please provide a username to initialize the bot. Got: ${JSON.stringify(username)}`)
     }
     if (!paperkey || typeof paperkey !== 'string') {
-      // Don't want to accidentally print the paperkey to STDERR
+      // Don't want to accidentally print the paperkey to STDERR.
       throw new Error(`Please provide a paperkey to initialize the bot.`)
     }
+    const homeDir = (options && options.homeDir) || randomTempDir()
+    this.homeDir = homeDir
 
-    await keybaseServiceStartup(this.homeDir)
-    await keybaseExec(this.homeDir, ['oneshot', '--username', username], {stdinBuffer: paperkey})
+    await keybaseServiceStartup(homeDir)
+    await keybaseExec(homeDir, ['oneshot', '--username', username], {stdinBuffer: paperkey})
 
     const currentDPair = await this._currentUsernameAndDevice()
     if (currentDPair && currentDPair.username && currentDPair.devicename) {
+      this.initialized = 'paperkey'
       this.username = currentDPair.username
       this.devicename = currentDPair.devicename
       this.verbose = options ? Boolean(options.verbose) : false
@@ -46,10 +51,17 @@ class BaseBot {
     }
   }
 
-  async initFromCurrentCredentials(homeDir: string, options?: InitOptions) {
-    this.homeDir = homeDir
+  /**
+   * Initialize a Keybase bot off of a currently running service.
+   * @memberof BaseBot
+   */
+  async initWithExistingService(options?: InitOptions) {
+    if (options && options.homeDir) {
+      this.homeDir = options.homeDir
+    }
     const currentDPair = await this._currentUsernameAndDevice()
     if (currentDPair && currentDPair.username && currentDPair.devicename) {
+      this.initialized = 'currentService'
       this.username = currentDPair.username
       this.devicename = currentDPair.devicename
       this.verbose = options ? Boolean(options.verbose) : false
@@ -63,9 +75,13 @@ class BaseBot {
    * @memberof BaseBot
    */
   async deinit(): Promise<void> {
-    await keybaseExec(this.homeDir, ['logout'])
-    await keybaseExec(this.homeDir, ['ctl', 'stop', '--shutdown'])
-    await rmdirRecursive(this.homeDir)
+    if (this.initialized === 'paperkey') {
+      await keybaseExec(this.homeDir, ['logout'])
+      await keybaseExec(this.homeDir, ['ctl', 'stop', '--shutdown'])
+      if (this.homeDir && typeof this.homeDir === 'string') {
+        await rmdirRecursive(this.homeDir)
+      }
+    }
   }
 
   /**
