+++ lib/chat/index.js
@@ -13,7 +13,7 @@ import type {
   ChatDeleteOptions,
 } from './types'
 
-export type ApiCommandArg = {|method: string, options: Object, homeDir: string|}
+export type ApiCommandArg = {|method: string, options: Object, homeDir?: string|}
 
 const runApiCommand = async (arg: ApiCommandArg): Promise<any> => {
   const options = formatOptions(arg.options)
