--- lib/entry.js
@@ -1,3 +1,0 @@
-// @flow
-import Bot from './bot.js'
-export default Bot
