+++ lib/index.js
@@ -12,7 +12,7 @@ class Bot extends BaseBot {
 
   async init(username: string, paperkey: string, options?: InitOptions) {
     await super.init(username, paperkey, options)
-    await this.chat.initFromCurrentCredentials(this.homeDir, options)
+    await this.chat.initWithExistingService({...options, homeDir: this.homeDir})
   }
 }
