+++ lib/utils/index.js
@@ -1,0 +1,6 @@
+// @flow
+export {default as formatOptions} from './formatOptions'
+export {default as keybaseExec} from './keybaseExec'
+export {default as keybaseServiceStartup} from './keybaseServiceStartup'
+export {default as randomTempDir} from './randomTempDir'
+export {default as rmdirRecursive} from './rmdirRecursive'
